+++ src/styles/globals.css
@@ -123,3 +123,15 @@
     @apply bg-background text-foreground;
   }
 }
+
+@layer utilities {
+  @keyframes heartbeat {
+    0% { transform: scale(1); }
+    30% { transform: scale(1.5); }
+    60% { transform: scale(0.9); }
+    100% { transform: scale(1); }
+  }
+  .animate-heartbeat {
+    animation: heartbeat 0.5s cubic-bezier(.4,1.6,.6,1) 1;
+  }
+}
